# internal/api/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  exercise.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  exercise.Equipment:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1/admin/equipment:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Equipment'
            type: array
      security:
      - BearerAuth: []
      summary: List all equipment
      tags:
      - equipment
  /api/v1/admin/equipment/{id}:
    get:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.Equipment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get equipment by ID
      tags:
      - equipment
  /api/v1/admin/exercise-categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Category'
            type: array
      security:
      - BearerAuth: []
      summary: List all exercise categories
      tags:
      - exercise-categories
  /api/v1/admin/exercise-categories/{id}:
    get:
      parameters:
      - description: Exercise Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get exercise category by ID
      tags:
      - exercise-categories
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Log in a user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: JWT-based authentication using Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"